scanner/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
discord.py==2.6.4
fonttools==4.60.1
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
httplib2==0.31.2
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pynput==1.8.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.22.0
yt-dlp==2025.11.12
